[Pal,LibOS] Disable sysfs topology feature by default
This commit disables sysfs topology feature by default as it is still
under development and may contain security vulnerabilities, and exposes
a new manifest option `fs.experimental__enable_sysfs_topology` to enable
this feature for testing purposes. This is a temporary change and the
feature will be re-enabled in future after thorough validation.

Co-authored-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>
Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>
Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>